Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.25.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.0
